Add modules for static
