+++ .github/workflows/modules/store.js
@@ -1,0 +1,36 @@
+import Store from './store.js';
+
+export default class UI {
+  static displayBooks() {
+    const books = Store.getBooks();
+    books.forEach((book) => UI.addBookToList(book));
+  }
+
+  static addBookToList(book) {
+    const books = document.querySelector('.books');
+    const newBook = document.createElement('div');
+    newBook.innerHTML = `
+        <div>
+        <div class="bookDiv">
+        <h4 id="title" class="">${book.title} &nbsp; &nbsp; &nbsp; &nbsp; by &nbsp; &nbsp; &nbsp; &nbsp; ${book.author}</h4>
+        <h4 class="hidy">${book.id}</h4>
+        <button class="delete">Remove</button>
+        </div>
+        <hr class="hr">
+        </div>
+        `;
+    newBook.classList.add('newBook');
+    books.appendChild(newBook);
+  }
+
+  static deleteBook(el) {
+    if (el.classList.contains('delete')) {
+      el.parentElement.parentElement.remove();
+    }
+  }
+
+  static clearFields() {
+    document.querySelector('.title').value = '';
+    document.querySelector('.author').value = '';
+  }
+}
